requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.6
django-environ==0.9.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-spectacular==0.25.1
flake8==6.0.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
pycodestyle==2.10.0
pyflakes==3.0.1
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
simplejson==3.18.3
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
